requirements.txt
streamlit
opencv-python
scikit-learn
scikit-image
Pillow
numpy
joblib
